pyproject.toml
[tool.poetry]
name = "btc-wallet"
version = "0.1.0"
description = "A very simple BTC wallet generator"
authors = ["Vlad Nedelcu <nedelcuvd@gmail.com>"]
license = "GNU"
readme = "README.md"
packages = [{include = "btc_wallet", from = "src"}]

[tool.poetry.dependencies]
python = "^3.12.2"
secp256k1 = "^0.14.0"
pycryptodome = "^3.20.0"
python-bitcoinlib = "^0.12.2"
pydantic = "^2.7.1"
ecdsa = "^0.19.0"
httpx = "^0.27.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"
flake8 = "^7.0.0"
black = "^24.4.2"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
